requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scipy>=1.7.0
tensorboard>=2.7.0
transformers>=4.15.0
diffusers>=0.10.0
accelerate>=0.12.0
gdown>=4.4.0
einops>=0.4.0
timm>=0.5.0
albumentations>=1.0.0
pytorch-lightning>=1.5.0
omegaconf>=2.1.0
hydra-core>=1.1.0 